Fix integration dependencies
